Offer fork in readme
Given the amount of changes in my fork, I don't really know if you want to merge that into your master given it doesn't use wget and it uses pThreads and cURL.That being said, it could be pointed to (particularly for those who need a Windows compatible version). The fork is much faster than the one here but does have additional dependencies as mentioned.